requirements.txt
ursina
sounddevice
numpy
openai-whisper
groq
gtts
pydub
scipy